Reset UI on WebGL context loss
On Emscripten, the WebGL context may be loss for any reason, for example
when restoring the application from the background, the GPU force-
refreshed the application to reclaim memory, etc. In these cases, canvas
data is lost and this commit accommodates that situation by reinitiali-
zing the UI when the browser has restored the context.